backend/weather_data_ingestion/requirements.txt
python-dotenv==1.0.1
requests>=2.31.0
pandas>=2.2.0
pydantic==2.6.1
pydantic-settings==2.1.0 